% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{add_geo_level}
\alias{add_geo_level}
\title{Add Geo Level}
\usage{
add_geo_level(dt)
}
\description{
Add geographical levels to the dataset. They are:
\itemize{
\item \code{L} for country (\code{Land})
\item \code{F} for county (\code{Fylke})
\item \code{K} for bigger municipality (\code{Kommune}) ie. population > 10,000
\item \code{km} for smaller municipality (\code{Kommune}) ie. population < 10,000
\item \code{B} for town (\code{Bydele})
}
}
\keyword{internal}
