% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{check_cube}
\alias{check_cube}
\alias{sjekk_kube}
\alias{cc}
\title{Check Outliers Raw File}
\usage{
check_cube(name = NULL, year = NULL, type = c("KH", "NH"), km = FALSE, ...)

sjekk_kube(name = NULL, year = NULL, type = c("KH", "NH"), km = FALSE, ...)

cc(name = NULL, year = NULL, type = c("KH", "NH"), km = FALSE, ...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Make funcion name standardize
}
\examples{
\dontrun{
check_cube("TRIVSEL")
# ->
check_outliers("TRIVSEL")
}
}
\keyword{internal}
