% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{check_file}
\alias{check_file}
\title{Check File}
\usage{
check_file(name = NULL, year = NULL, type = c("KH", "NH"), ...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{check_file()} was renamed to \code{check_cube()} to be more explicit
}
\keyword{internal}
