% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{check_outliers}
\alias{check_outliers}
\alias{sjekk_outlier}
\alias{co}
\title{Check Outliers Raw File}
\usage{
check_outliers(name = NULL, year = NULL, type = c("KH", "NH"), km = FALSE, ...)

sjekk_outlier(name = NULL, year = NULL, type = c("KH", "NH"), km = FALSE, ...)

co(name = NULL, year = NULL, type = c("KH", "NH"), km = FALSE, ...)
}
\arguments{
\item{name}{Filename of KUBE raw \code{csv} file.}

\item{year}{Which year of the \code{KHxxxxNESSTAR} folder to select where \code{xxxx}
is the selected year. Default is NULL to select the current year as in
options \code{kh.year}.}

\item{type}{Type of folder of either \code{KH} or \code{NH}. Default is \code{KH}.}

\item{km}{Logical value. To include or exclude small municipalities. Default
is to exclude them ie. \code{km = FALSE}}

\item{...}{Additional arguments}
}
\description{
Check outliers in raw kube files
}
\examples{
\dontrun{
dt <- check_outliers("TRIVSEL", year = 2022)
}
}
