% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{count_pop}
\alias{count_pop}
\title{Count Municipalities Population}
\usage{
count_pop(name = "BEFOLK_GK", dir = "current", overwrite = FALSE)
}
\arguments{
\item{name}{Population filename with standard name starts with \code{BEFOLK_GK}}

\item{dir}{Select directory between \code{current} or \code{previous} year}

\item{overwrite}{Overwrite existing \code{BEF-Kommune-xxxx.rds} file}
}
\description{
Create a dataset separating big and small municipalities based
on their number of population. The cutoff is 10,000 population. Capital
letter \code{K} denotes big municipalities while small letter \code{k} for small
municipalities. The created file will be called \code{BEF-Kommune-xxxx} where
\code{xxxx} represent the selected year.
}
