% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{create_pop_ref}
\alias{create_pop_ref}
\title{Create Population Reference}
\usage{
create_pop_ref(name = "BEFOLK_GK", year = NULL, overwrite = FALSE)
}
\arguments{
\item{name}{Population filename with standard name starts with \code{BEFOLK_GK}}

\item{year}{Which year of the \code{KHxxxxNESSTAR} folder to select where \code{xxxx}
is the selected year. Default is NULL to select the current year as in
options \code{kh.year}.}

\item{overwrite}{Overwrite existing \code{BigSmall-Kommuner-REF-xxxx.rds} file}
}
\description{
Create a dataset separating big and small municipalities based
on their number of population. The cutoff is 10,000 population. Capital
letter \code{K} denotes big municipalities while small letter \code{k} for small
municipalities. The created file will be called
\code{BigSmall-Kommuner-REF-xxxx.rds} where \code{xxxx} represent the selected year.
}
