% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{diff_change}
\alias{diff_change}
\title{Get the Difference}
\usage{
diff_change(dt, dim, ...)
}
\arguments{
\item{dt}{Dataset}

\item{dim}{Dimension dataset produced by \code{get_grid()}}

\item{...}{Additional arguments}
}
\description{
Get the difference for change in the cube measure variables ie.
the varibles that are created when running the function \code{LagKUBE()} from
\code{KHfunctions}.
}
