% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\docType{data}
\name{.env_dim}
\alias{.env_dim}
\title{Env for Dimensions}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.env_dim
}
\description{
Where all object for dimensions will be kept.
}
\keyword{datasets}
