% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\docType{data}
\name{.env_key}
\alias{.env_key}
\title{Env for Key Variables}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.env_key
}
\description{
Where key variables for dataset will be kept
}
\keyword{datasets}
