% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\docType{data}
\name{.env_misc}
\alias{.env_misc}
\title{Env for Misc Variables}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.env_misc
}
\description{
When there are a need for it
}
\keyword{datasets}
