% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{get_dir}
\alias{get_dir}
\title{Get Directory}
\usage{
get_dir(year = NULL, os = OS)
}
\arguments{
\item{year}{Which year of the \code{KHxxxxNESSTAR} folder to select where \code{xxxx}
is the selected year. Default is NULL to select the current year as in
options \code{kh.year}.}

\item{os}{Operating system}
}
\description{
Get the directory where the files will be read from.
}
