% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{get_dir_ref}
\alias{get_dir_ref}
\title{Get Reference Directory}
\usage{
get_dir_ref(year = NULL, os = OS)
}
\arguments{
\item{year}{Which year to refer to. Default is option \code{kh.year}}

\item{os}{Operating system}
}
\description{
Get the directory where reference files will be kept. These
files include population reference files and database files.
}
