% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension.R
\name{get_grid}
\alias{get_grid}
\title{Get Grid of All Dimensions}
\usage{
get_grid(dt, vars)
}
\arguments{
\item{dt}{Dataset}

\item{vars}{Key variables ie. the output from \code{get_key()}}
}
\description{
Get cross join of all the dimensions for the analysis in the
dataset.
}
