% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHcompare-package.R
\docType{package}
\name{KHompare-package}
\alias{KHompare}
\alias{KHompare-package}
\title{KHompare: Compare Cubes Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Compare and identify abnormal yearly changes for KHelse cube data.
}
\author{
\strong{Maintainer}: Yusman Kamaleri \email{ybka@fhi.no} (\href{https://orcid.org/0000-0001-5014-3665}{ORCID})

}
\keyword{internal}
