% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cube}
\alias{plot_cube}
\alias{pc}
\title{Plot Data}
\usage{
plot_cube(data, geo, var, value = c("pct", "num", "raw"), interactive = TRUE)

pc(data, geo, var, value = c("pct", "num", "raw"), interactive = TRUE)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Make funcion name standardize
}
\examples{
\dontrun{
plot_cube(dt, geo = 3, var = "TELLER")
# ->
plot_outliers(dt, geo = 3, var = "TELLER")
}
}
\keyword{internal}
