% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_outliers}
\alias{plot_outliers}
\alias{po}
\title{Plot Data}
\usage{
plot_outliers(
  data,
  geo,
  var,
  value = c("pct", "num", "raw"),
  interactive = TRUE
)

po(data, geo, var, value = c("pct", "num", "raw"), interactive = TRUE)
}
\arguments{
\item{data}{Cube dataset from \code{check_cube()} function}

\item{geo}{Geographical code}

\item{var}{Cube measure variable}

\item{value}{Percentage or numeric change. Select either \code{pct} or \code{num}}

\item{interactive}{Logical value. Interactive plot or static}
}
\description{
Plot cube dataset for selected geo code and cube measurement
variable
}
\examples{
\dontrun{
plot_outliers(dt, geo = 3, var = "TELLER")
}
}
