% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{read_cube}
\alias{read_cube}
\alias{les_kube}
\title{Read Raw File}
\usage{
read_cube(file = NULL, ...)

les_kube(file = NULL, ...)
}
\arguments{
\item{file}{Raw \code{csv} file. Accept \code{KUBE} name if it's unique}

\item{...}{Additional arguments}
}
\description{
Read raw kube files
}
\examples{
\dontrun{
dt <- read_cube("REGNFERD", dir = "current")
}
}
