% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{read_file}
\alias{read_file}
\title{Read Raw File}
\usage{
read_file(file = NULL, ...)
}
\arguments{
\item{file}{Raw \code{csv} file. Accept \code{KUBE} name if it's unique}

\item{...}{Additional arguments}
}
\description{
Read raw kube files
}
\examples{
\dontrun{
dt <- read_file("REGNFERD", dir = "current")
}
}
