% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{check_cube}
\alias{check_cube}
\alias{sjekk_kube}
\title{Check Raw File}
\usage{
check_cube(file = NULL, ...)

sjekk_kube(file = NULL, ...)
}
\arguments{
\item{file}{Raw \code{csv} file. Accept \code{KUBE} name if it's unique}

\item{...}{Additional arguments}
}
\description{
Check raw kube files
}
\examples{
\dontrun{
dt <- check_cube("REGNFERD", dir = "current")
}
}
