% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{get_dir}
\alias{get_dir}
\title{Get Directory}
\usage{
get_dir(dir = c("current", "previous"), os = OS)
}
\arguments{
\item{dir}{Select directory between \code{current} or \code{previous} year}

\item{os}{Operating system}
}
\description{
Get the directory where the files will be read from.
}
