% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cube}
\alias{plot_cube}
\alias{pc}
\title{Plot Data}
\usage{
plot_cube(data, geo, var, value = c("pct", "num"))

pc(data, geo, var, value = c("pct", "num"))
}
\arguments{
\item{data}{Cube dataset from \code{check_cube()} function}

\item{geo}{Geographical code}

\item{var}{Cube measure variable}

\item{value}{Percentage or numeric change. Select either \code{pct} or \code{num}}
}
\description{
Plot cube dataset for selected geo code and cube measurement
variable
}
\examples{
\dontrun{
plot_cube(dt, geo = 3, var = "TELLER")
}
}
