% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{read_befolk}
\alias{read_befolk}
\title{Big and Small Municipalities}
\usage{
read_befolk(name = "BEFOLK_GK", year = getOption("kh.year"), overwrite = FALSE)
}
\arguments{
\item{name}{Population filename with standard name starts with \code{BEFOLK_GK}}

\item{year}{Year for selection of population. Default is using global options \code{kh.year}}

\item{overwrite}{Overwrite existing \code{BEF-Kommune-xxxx.rds} file}
}
\description{
Create a dataset seperating big and small municipalities. The
cutoff is 10,000 population. Capital letter \code{K} denotes big municipalities while small letter
\code{k} for small municipalities.
}
