% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view_outliers}
\alias{view_outliers}
\alias{vis_uteligger}
\alias{vo}
\title{View Outliers}
\usage{
view_outliers(dt = NULL, var = NULL, nrows = NULL, levels = NULL)

vis_uteligger(dt = NULL, var = NULL, nrows = NULL, levels = NULL)

vo(dt = NULL, var = NULL, nrows = NULL, levels = NULL)
}
\arguments{
\item{dt}{Dataset object from \code{check_cube()} function}

\item{var}{Measure variables eg. RATE, MEIS etc}

\item{nrows}{Number of rows or select rows to show eg. \code{nrows = 10} or \code{nrows = 5:15}}

\item{levels}{Level of geographical granularity as in \code{getOption("kh.geo.levels")}}
}
\description{
View outliers for selected variables
}
\examples{
\dontrun{
 dt <- check_cube("ALKOHOL")
 view_outliers(dt, "RATE")
}
}
