% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view_outliers}
\alias{view_outliers}
\alias{vis_uteligger}
\title{View Outliers}
\usage{
view_outliers(dt = NULL, var = NULL, nrow = NULL)

vis_uteligger(dt = NULL, var = NULL, nrow = NULL)
}
\arguments{
\item{dt}{Dataset object from \code{check_cube()} function}

\item{var}{Measure variables eg. RATE, MEIS etc}

\item{nrow}{Number of rows to show}
}
\description{
View outliers for selected variables
}
\examples{
\dontrun{
 dt <- check_cube("ALKOHOL")
 view_outliers(dt, "RATE")
}
}
